+++ packages/core/src/Console.ts
@@ -30,15 +30,17 @@ export default class Console extends Emitter {
 
   outputQueue: Output[] = [];
 
-  renderTimer: NodeJS.Timer | null = null;
-
-  restoreCursorOnExit: boolean = false;
-
-  started: boolean = false;
-
-  stopped: boolean = false;
-
   tool: Tool<any>;
+
+  protected final: boolean = false;
+
+  protected renderTimer: NodeJS.Timer | null = null;
+
+  protected restoreCursorOnExit: boolean = false;
+
+  protected started: boolean = false;
+
+  protected stopped: boolean = false;
 
   private writers: typeof BOUND_WRITERS;
 
@@ -148,7 +150,7 @@ export default class Console extends Emitter {
   handleSignal = () => {
     this.start();
     this.debug('SIGINT or SIGTERM handled');
-    this.stop('Process has been terminated.', true);
+    this.stop(new Error('Process has been terminated.'), true);
   };
 
   /**
@@ -167,6 +169,13 @@ export default class Console extends Emitter {
     }
 
     return this;
+  }
+
+  /**
+   * Return true if the final render.
+   */
+  isFinal(): boolean {
+    return this.final;
   }
 
   /**
@@ -233,6 +242,7 @@ export default class Console extends Emitter {
    */
   renderFinalOutput(error: Error | null) {
     this.debug('Rendering final console output');
+    this.final = true;
 
     // Stop the render loop
     this.stopRenderLoop();
@@ -313,15 +323,9 @@ export default class Console extends Emitter {
   /**
    * Stop the console rendering process.
    */
-  stop(message: string | Error | null = null, force: boolean = false) {
+  stop(error: Error | null = null, force: boolean = false) {
     if (this.stopped) {
       return;
-    }
-
-    let error = null;
-
-    if (message !== null) {
-      error = message instanceof Error ? message : new Error(message);
     }
 
     if (error) {