+++ packages/core/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Added `Tool#isCI` method to check if in a CI environment.
 - Added `Tool#isPluginEnabled` method to check if a plugin by type has been enabled.
+- Added `Console#isFinal` method to check whether it's the final render or not.
 - Added `Plugin#blueprint`, `Reporter#blueprint`, and `Routine#blueprint` as a means to validate and
   build options passed to the constructor.
 - Added `Reporter#hasColorSupport` to check for terminal color support.
@@ -13,6 +14,7 @@
 
 - Deprecated `Tool#loadWorkspacePackages`. Use `Tool#getWorkspacePackages` instead.
 - TS: `Tool#loadConfig`, `loadPlugins`, `loadReporters` have been marked `protected`.
+- TS: `Console#final`, `started`, `stopped` have been marked `protected`.
 
 # 1.4.1 - 2019-01-08